+++ ChangeLog.md
@@ -1,9 +1,12 @@
 # CHANGELOG CLIENJOYHOLIDAYSV2 FOR [DOLIBARR ERP CRM](https://www.dolibarr.org)
 
+## 2.0
+- NEW : Mise en place d'un système de tarif par défaut global si aucuns pays n'est renseignés modifiable 
+via l'interface de paramétrage du module.
+ Systeme de tarification en fonction du pays de destination modifiable dans un dictionnaire - 01/02/2023- 2.0.0
 
 ## 1.0
 - NEW : Ajout des images et des traductions.
-Ajout de la page de création des formules de voyages. 
-Création de la liste des formules de voyages - 01/02/2023 - 1.0.0
-
+  Ajout de la page de création des formules de voyages.
+  Création de la liste des formules de voyages - 01/02/2023 - 1.0.0